Problems to upload csv file
I have quit some problems to import the csv file. I have tried to use the fetch as well as installing and importing the parsers such as d3 and CSVtoJSON. I have tried to do this both directly in the class component in componentDidMount and in a separate component with useEffect. i have even tried to use fetch and then get the data in plain text and then work from there but it das all not work.

Have I misunderstood something or can you not use react Hooks or fetch when you also use reactRoutes? or is it more likely that something went wrong with the install of create-react-app, d3, or CSVtoJSON? Here i get the error that I can not find "de" or "fs even though the they are installed and importet.

But even then I would expect the fetch to be working. Here I do not get an actual error but it passes in an Html element / Html code which I also do not understand. 

I would appreciate to know if I am on the right track and a possible solution to this problem.

you can find my code here. sorry if it is all confusing but I have tried so many combinations to solve this.